refactor: conditional rendering CVC, and types to expiryDate
