+++ src/components/loaders/CircularLoaderWithTimer.tsx
@@ -3,7 +3,7 @@ type Props = {
   size?: number;
 };
 
-export default function CountdownCircleTimer(
+export default function CircularLoaderWidthTimer(
   {
     duration = 10,
     size = 20,
