+++ src/components/modals/ModalCardInformation.tsx
@@ -5,25 +5,22 @@ import {
 } from '@dynamic-framework/ui-react';
 import { useTranslation } from 'react-i18next';
 
-import useCVC from '../../hooks/useCVC';
+import { API_ACCOUNT_LIST_FILTER } from '../../config/widgetConfig';
 import { Account } from '../../services/interface';
 import { useAppSelector } from '../../store/hooks';
 import { getAccountSelected } from '../../store/selectors';
 import CardItem from '../CardItem';
 import ItemCardInfo from '../ItemCardInfo';
-import CircularLoaderWidthTimer from '../loaders/CircularLoaderWithTimer';
+import ItemCardInfoCVC from '../ItemCardInfoCVC';
 import NumberCardInfo from '../NumberCardInfo';
-
-const DURATION = 20;
 
 export default function ModalCardInformation() {
   const { t } = useTranslation();
   const { closePortal } = useDPortalContext();
   const account = useAppSelector(getAccountSelected) as Account;
-  const cvc = useCVC(DURATION);
 
   return (
-    <DModal name="modalOTP" centered staticBackdrop size="lg">
+    <DModal name="modalCardInformation" centered staticBackdrop size="lg">
       <DModal.Header onClose={closePortal} showCloseButton>
         <h5 className="fw-bold">{t('cardInfo.title')}</h5>
       </DModal.Header>
@@ -47,19 +44,7 @@ export default function ModalCardInformation() {
               name={t('cardInfo.cardNumber')}
               value={<NumberCardInfo cardNumber={account.accountNumber} />}
             />
-            <ItemCardInfo
-              name={t('cardInfo.validUntil')}
-              value={account?.accountNumber}
-            />
-            <ItemCardInfo name={t('cardInfo.cvc')} value={cvc.cvc} />
-            <div className="d-flex gap-2">
-              <CircularLoaderWidthTimer size={20} duration={DURATION} />
-              <p className="text-gray-500">
-                {`${t('cardInfo.cvcWillChange')} ${cvc.secondsLeft} ${t(
-                  'cardInfo.seconds',
-                )}`}
-              </p>
-            </div>
+            {API_ACCOUNT_LIST_FILTER === 'credit-card' && <ItemCardInfoCVC expiryDate={account.expiryDate} />}
           </div>
         </div>
       </DModal.Body>
